refactor(rebalancer): use cli logger tooling
